chore: 재귀함수 테스트 코드 삭제, 파일 분리
